+++ src/main/kotlin/com/team2/server/party/controller/PartyInviteController.kt
@@ -3,7 +3,7 @@ package com.team2.server.party.controller
 import com.team2.server.auth.principal.UserPrincipal
 import com.team2.server.common.web.ApiResponse
 import com.team2.server.party.dto.ActivateInviteLinkResponse
-import com.team2.server.party.service.PartyInviteService
+import com.team2.server.party.application.service.PartyInviteService
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
